requirements.txt
requests==2.25.1
streamlit==0.76.0
Pillow==8.1.0
